+++ NEWS.md
@@ -1,3 +1,27 @@
+## Version 1.1.0
+
+This is a major release, resolving some problems with `plot_discrim()` and adding considerable functionality
+
+* Fixed buglet in `plot_discrim()` where the variables in a formula needed to be reversed.
+* `plot_discrim()` now gets the `data` from the object, if not supplied.
+* Added `tile.alpha` arg to `plot_discim()`; document how to customize colors, shapes, ...
+* `plot_discrim()` gains an `ellipse` argument to draw data ellipses for the groups.
+* Added `ellipse.args` to control stat_ellipse() parameters
+* Added `labels` and `labels.args` for class labels at group means
+* Flesh out package description in README.
+* Default for `posterior` in `predict_discrim()` becomes `FALSE`, as rarely needed
+* `plot_discrim()` now allows plotting the results in discriminant space, e.g., `LD2 ~ LD1`
+* It also generates nice labels for the discriminant dimensions
+* Added `scores.lda()` to extract discriminant scores
+* Fixed problems in reflect()
+* Added examples to README
+* Started to add `testthat` to package checks
+* Added another vignette using the `MASS::painters` data
+* Added `painters2` dataset
+* Added `datasets.csv` file to allow use of `@concept` tags in pkgdown site
+* Added `rev.axes` to reverse axes in `plot_discrim()`, useful for plots in discrim space, where  orientation is arbitrary
+* Added `xlim`, `ylim` arguments to control axis limits; useful for plots in discrim space.
+
 ## Version 1.0.0
 
 This is a major release, adding methods for linear/quadratic discriminant analysis and improving package documentation